+++ src/main/java/fll/db/ImportDB.java
@@ -675,6 +675,11 @@ public final class ImportDB {
       upgrade35To36(connection);
     }
 
+    dbVersion = Queries.getDatabaseVersion(connection);
+    if (dbVersion < 37) {
+      upgrade36To37(connection);
+    }
+
     // NOTE: when adding new tournament parameters they need to be explicitly set in
     // importTournamentParameters
 
@@ -1349,8 +1354,23 @@ public final class ImportDB {
           + GenerateDB.PERFORMANCE_TABLE_NAME
           + " SET tablename = 'UNKNOWN'");
     }
-    
+
     setDBVersion(connection, 36);
+  }
+
+  /**
+   * Remove judging_station from schedule table.
+   */
+  private static void upgrade36To37(final Connection connection) throws SQLException {
+    LOGGER.debug("Upgrading database from 36 to 37");
+
+    try (Statement stmt = connection.createStatement()) {
+      if (checkForColumnInTable(connection, "schedule", "judging_station")) {
+        stmt.executeUpdate("ALTER TABLE schedule DROP COLUMN judging_station");
+      }
+    }
+
+    setDBVersion(connection, 37);
   }
 
   /**
@@ -1962,11 +1982,11 @@ public final class ImportDB {
       destPrep.executeUpdate();
     }
 
-    try (PreparedStatement sourcePrep = sourceConnection.prepareStatement("SELECT team_number, judging_station" //
+    try (PreparedStatement sourcePrep = sourceConnection.prepareStatement("SELECT team_number" //
         + " FROM schedule WHERE tournament=?");
         PreparedStatement destPrep = destinationConnection.prepareStatement("INSERT INTO schedule" //
-            + " (tournament, team_number, judging_station)" //
-            + " VALUES (?, ?, ?)")) {
+            + " (tournament, team_number)" //
+            + " VALUES (?, ?)")) {
 
       sourcePrep.setInt(1, sourceTournamentID);